refactor translation dialog activation to main form ensure translate all link types' coordinates
