+++ src/ui/EPANET/frmMainEPANET.py
@@ -39,7 +39,6 @@ from ui.EPANET.frmEnergyReport import frmEnergyReport
 from ui.EPANET.frmCalibrationData import frmCalibrationData
 from ui.EPANET.frmCalibrationReportOptions import frmCalibrationReportOptions
 from ui.frmGenericPropertyEditor import frmGenericPropertyEditor
-from ui.frmTranslateCoordinates import frmTranslateCoordinates
 
 from core.epanet.epanet_project import EpanetProject as Project
 from core.epanet.inp_reader_project import ProjectReader
@@ -136,7 +135,6 @@ class frmMainEPANET(frmMain):
         frmMain.__init__(self, q_application)
         self.on_load(tree_top_item_list=self.tree_top_items)
         self.project_settings = DefaultsEPANET("", self.project)
-        self.translate_coord_dialog = None
         self.tree_types = {
             self.tree_Patterns[0]: Pattern,
             self.tree_Curves[0]: Curve,
@@ -630,20 +628,6 @@ class frmMainEPANET(frmMain):
         from ui.EPANET.frmFind import frmFind
         frmF = frmFind(self, self.project)
         frmF.show()
-
-    def open_translate_coord_dialog(self, pt_src_ll, pt_src_ur):
-        # translate EPANET coords
-        self.setQgsMapToolTranslateCoords()
-        if not self.translate_coord_dialog:
-            self.translate_coord_dialog = frmTranslateCoordinates(self, pt_src_ll, pt_src_ur)
-        else:
-            self.translate_coord_dialog.set_coords_source_ll(pt_src_ll)
-            self.translate_coord_dialog.set_coords_source_ur(pt_src_ur)
-
-        self.translate_coord_dialog.show()
-        # result = frm.exec_()
-        # if result:
-        #     pass
 
     def edit_simulation_options(self):
         self.show_edit_window(self.get_editor('Hydraulics'))
